+++ env_var.sh
@@ -3,3 +3,7 @@
 export contact_name='Ashley Nylen'
 export contact_email='ashley.nylen@state.co.us'
 export issuing_organization='CDOT'
+
+# $env:contact_name='Ashley Nylen'
+# $env:contact_email='ashley.nylen@state.co.us'
+# $env:issuing_organization='CDOT'