+++ tests/tools/polygon_tools_test.py
@@ -1,4 +1,4 @@
-from translator.tools import polygon_tools
+from wzdx.tools import polygon_tools
 
 
 # --------------------------------------------------------------------------------Unit test for get_road_direction function--------------------------------------------------------------------------------