+++ wzdx/standard_to_enhanced/__init__.py
@@ -1,0 +1,3 @@
+from wzdx.standard_to_enhanced.cotrip_translator import wzdx_creator
+from wzdx.standard_to_enhanced.icone_translator import wzdx_creator
+from wzdx.standard_to_enhanced.navjoy_translator import wzdx_creator