+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+10/8/2021
+- Created python package from project
+- Creating iCone 
+
 7/22/2021
 - Created a tools directory (at root of translator) containing all shared code and all code that was not strictly related to a translator (general polygon and datetime manipulation)
 - Removed source_code directory and placed all contained files at the root of translator
@@ -15,4 +19,4 @@
 - Updated date_tools with new iso8601 parsing functions
 - Updated unit tests throughout project, coverage is currently at 87%
 - Tested all translators
-- TODO: Navjoy 568 Translator: Support additional zones per message (streetNameFrom2, ...)
+- [COMPLETED] TODO: Navjoy 568 Translator: Support additional zones per message (streetNameFrom2, ...)